+++ Assets/Scripts/Towers/Damage/MissileDamage.cs
@@ -2,38 +2,16 @@ using System;
 using UnityEngine;
 using AudioSystem;
 
-public class MissileDamage : MonoBehaviour, IDamageMethod
+public class MissileDamage : TowerDamage
 {
     public Animator cannonAnimator;
     public LayerMask enemiesLayer;
     [SerializeField] private Transform towerPivot;
 
-    GameManager gameManager;
-    [SerializeField] AudioData audioData;
     [SerializeField] private GameObject missile;
-    [NonSerialized] public float damage;
     public float explosionRadius;
-    private float fireRate;
-    private float delay;
-    public void Init(float damage, float fireRate)
-    {
-        gameManager = GameManager.Instance;
-        this.damage = damage;
-        this.fireRate = fireRate;
-        delay = 1f / fireRate;
-    }
 
-    public void UpdateDamage(float damage)
-    {
-        this.damage = damage;
-    }
-
-    public void UpdateFireRate(float fireRate)
-    {
-        this.fireRate = fireRate;
-    }
-
-    public void damageTick(Enemy target)
+    public override void DamageTick(Enemy target)
     {
         if (target)
         {
